+++ src/test/java/ru/vyarus/guice/ext/generator/support/bad/WrongUsageBean.java
@@ -4,6 +4,8 @@ import com.google.inject.ProvidedBy;
 import com.google.inject.internal.DynamicClassProvider;
 
 /**
+ * Class generation can't be used with non abstract types
+ *
  * @author Vyacheslav Rusakov
  * @since 10.12.2014
  */