+++ src/test/java/ru/vyarus/guice/ext/postprocess/support/AbstractBean.java
@@ -1,0 +1,13 @@
+package ru.vyarus.guice.ext.postprocess.support;
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 06.01.2015
+ */
+public abstract class AbstractBean {
+    public int called;
+
+    public void call() {
+        called++;
+    }
+}