+++ src/test/java/ru/vyarus/guice/ext/postprocess/support/Bean3.java
@@ -1,0 +1,8 @@
+package ru.vyarus.guice.ext.postprocess.support;
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 06.01.2015
+ */
+public class Bean3 {
+}